+++ functions.php
@@ -138,7 +138,7 @@ function display_author_info_conditionally() {
 	}
 }
 
-function display_series_author( $term = null, $echo = true ) {
+function display_volumes_author( $term = null, $echo = true ) {
     if ( ! $term ) {
         $term = get_queried_object();
     }
@@ -162,7 +162,7 @@ function display_series_author( $term = null, $echo = true ) {
     $author_name = esc_html( $author->display_name );
     $author_url  = esc_url( get_author_posts_url( $author_id ) );
 
-    $output  = '<div class="series-author">';
+    $output  = '<div class="volumes-author">';
     $output .= ' <a href="' . $author_url . '">' . $author_name . '</a>';
     $output .= '</div>';
 
@@ -174,12 +174,12 @@ function display_series_author( $term = null, $echo = true ) {
     return $output;
 }
 
-function display_series_name() {
-	$series_terms = get_the_terms(get_the_ID(), 'series');
-	if ($series_terms && !is_wp_error($series_terms)) {
-		$series_term = array_shift($series_terms);
-		$series_link = get_term_link($series_term);
-		echo '<p class="series-link"><a href="' . esc_url($series_link) . '">' . esc_html($series_term->name) . '</a></p>';
+function display_volumes_name() {
+	$volumes_terms = get_the_terms(get_the_ID(), 'volumes');
+	if ($volumes_terms && !is_wp_error($volumes_terms)) {
+		$volumes_term = array_shift($volumes_terms);
+		$volumes_link = get_term_link($volumes_term);
+		echo '<p class="volumes-link"><a href="' . esc_url($volumes_link) . '">' . esc_html($volumes_term->name) . '</a></p>';
 	}
 }