+++ patterns/template-taxonomy-volumes.php
@@ -3,8 +3,8 @@ $term = get_queried_object(); // Ottiene il termine attuale della tassonomia
 ?>
 
 <div class="wp-group">
-	<?php display_series_author(); ?>
-	<h1 class="series-title"><?php echo esc_html($term->name); ?></h1>
+	<?php display_volumes_author(); ?>
+	<h1 class="volumes-title"><?php echo esc_html($term->name); ?></h1>
 </div>
 
 <?php
@@ -12,7 +12,7 @@ $args = array(
 	'post_type' => array('post'), // Includi entrambi i tipi di post
 	'tax_query' => array(
 		array(
-			'taxonomy' => 'series',
+			'taxonomy' => 'volumes',
 			'field'    => 'slug',
 			'terms'    => $term->slug,
 		),
@@ -26,10 +26,10 @@ $args = array(
 $query = new WP_Query($args);
 ?>
 
-<div class="series-chapters">
+<div class="volumes-chapters">
 <?php
 if ($query->have_posts()) :
-	echo '<ul class="series-posts">';
+	echo '<ul class="volumes-posts">';
 	while ($query->have_posts()) : $query->the_post(); ?>
 		<li>
 			<h2 class="chapter-title"><a href="<?php the_permalink(); ?>"><?php the_title(); ?></a></h2>
@@ -39,6 +39,6 @@ if ($query->have_posts()) :
 
 	wp_reset_postdata();
 else :
-	echo '<p>Nessun contenuto disponibile per questa serie.</p>';
+	echo '<p>Nessun contenuto disponibile per questo volume.</p>';
 endif;
 ?>