+++ taxonomy-volumes.php
@@ -2,7 +2,7 @@
 
 <div id="primary" class="content-area">
 	<main id="main" class="site-main" role="main">
-		<?php get_template_part('patterns/template-taxonomy-series'); ?>
+		<?php get_template_part('patterns/template-taxonomy-volumes'); ?>
 	</main>
 </div>