+++ CHANGELOG.md
@@ -5,10 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.3.0] 2025-07-10
 ### Added
+- Adds Volumes and Author's List pages
+- Updates the Header adding a Context Menu to browser content by Volumes and Authors
 - Displays the serie's Author
 ### Changed
+- Taxonomy "series" has been renamed "volumes"
 - Hides the pagination box in `patterns/template-query-loop.php` when no pages are available
 - `more-link` bottom margin changed for better post spacing
 
@@ -42,8 +45,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial `index.php`, `style.css`, and `functions.php`
 - This Changelog
 
-[unreleased]: https://github.com/erremauro/cigno-zen/compare/v0.2.1...HEAD
-[0.2.0]: https://github.com/erremauro/cigno-zen/releases/tag/v0.2.1
+[unreleased]: https://github.com/erremauro/cigno-zen/compare/v0.3.0...HEAD
+[0.30.0]: https://github.com/erremauro/cigno-zen/releases/tag/v0.3.0
+[0.2.1]: https://github.com/erremauro/cigno-zen/releases/tag/v0.2.1
 [0.2.0]: https://github.com/erremauro/cigno-zen/releases/tag/v0.2.0
 [0.1.0]: https://github.com/erremauro/cigno-zen/releases/tag/v0.1.0
 [0.0.1]: https://github.com/erremauro/cigno-zen/releases/tag/v0.0.1